requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
tortoise-orm==0.20.0
aerich==0.7.2
asyncpg==0.29.0
python-multipart==0.0.6
python-decouple==3.8
pydantic==2.5.3
pydantic-settings==2.1.0
pillow==10.2.0
aiofiles==23.2.1
python-dotenv>=1.0.0
argon2-cffi>=23.1.0
passlib>=1.7.4
psycopg2-binary>=2.9.9
httpx
PySocks
